@@ -1,50 +1,27 @@
 {
-  "name": "angular",
-  "version": "0.0.0",
+  "name": "angular-starter19",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.0.8",
-    "@angular/common": "^11.0.8",
-    "@angular/compiler": "^11.0.8",
-    "@angular/core": "^11.0.8",
-    "@angular/forms": "^11.0.8",
-    "@angular/platform-browser": "^11.0.8",
-    "@angular/platform-browser-dynamic": "^11.0.8",
-    "@angular/router": "^11.0.8",
+    "@angular/animations": "19.2.17",
+    "@angular/common": "19.2.17",
+    "@angular/compiler": "19.2.17",
+    "@angular/core": "19.2.17",
+    "@angular/forms": "19.2.17",
+    "@angular/platform-browser": "19.2.17",
+    "@angular/platform-browser-dynamic": "19.2.17",
+    "@angular/router": "19.2.17",
     "@popperjs/core": "^2.11.8",
     "bootstrap": "^5.3.8",
-    "popperjs/core": "*",
-    "popperjs/coree": "*",
-    "rxjs": "^6.6.3",
-    "sweetalert2": "^11.26.10",
-    "tslib": "^2.1.0",
-    "zone.js": "^0.11.3"
-  },
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "firebase": "11.1.0",
+    "rxjs": "7.8.2",
+    "sweetalert2": "11.26.17",
+    "tslib": "2.8.1",
+    "zone.js": "0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.4",
-    "@angular/cli": "~11.0.4",
-    "@angular/compiler-cli": "~11.0.4",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "@angular-devkit/build-angular": "^19.0.0",
+    "@angular/cli": "^19.0.0",
+    "@angular/compiler-cli": "^19.0.0",
+    "typescript": "~5.6.0"
   }
 }
