@@ -2,18 +2,24 @@ import { Component, OnInit } from '@angular/core';
 import { ProductoInterface } from '../../../models/producto.interface';
 import { ProductoGestionService } from '../../../services/producto-gestion.service';
 import { SweetAlertService } from '../../../services/sweet-alert.service';
+import { Router } from '@angular/router';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-lista-detalle-producto',
+  standalone: false, // <-- Cambia de true a false
   templateUrl: './lista-detalle-producto.component.html',
   styleUrls: ['./lista-detalle-producto.component.css'],
 })
 export class ListaDetalleProductoComponent implements OnInit {
-  ProductoList: ProductoInterface[];
+  ProductoList$: Observable<ProductoInterface[]>;
   constructor(
     private prodGestSer: ProductoGestionService,
-    private alertService: SweetAlertService
+    private alertService: SweetAlertService,
+    private rutas:Router,
   ) {
+this.ProductoList$=this.prodGestSer.obtenerListaProductos$;
+
     prodGestSer.detalleProductoEmitter.subscribe(
       (productoEmitido: ProductoInterface) => {
         this.alertService.mostrarExito(
@@ -25,10 +31,13 @@ export class ListaDetalleProductoComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.ProductoList = this.prodGestSer.ProductoListService;
+   
+    this.prodGestSer.cargarProductosDb();
+
   }
 
-  // onAgregarProductoList (Evento: ProductoInterface){
-  //   this.prodGestSer.onAgregarProductoListService(Evento);
-  // }
+  agregagarProducto(){
+this.rutas.navigate(['/AgregarProducto']);
+
+  }
 }
